✨ `utilities` directory for behind-the-scenes tools
Add `utilities` to `_config.yml` exclude list
Add `utilities` directory for website-building utilities
Add `overlay_images` utility for generating splash images

Additional tools are sometimes needed that can be helpful for
maintaining site, but should not be published as standalone tools.
`utilities/overlay_images` is the tool for generating the splash
images that are stored in `assets/images/splash_img/` and
heavily featured on the software pages.